% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvalingam-c.R
\name{vis_indices}
\alias{vis_indices}
\title{Line simplification using Visvalingam's algorithm}
\usage{
vis_indices(x, y, n)
}
\arguments{
\item{x}{points}

\item{y}{points}

\item{n}{number of points to keep}
}
\value{
logical vector giving location of the \code{n} simplified indices
}
\description{
Visvalingam's algorithm iteratively deletes vertices from a line based up
the \code{effecive area} of the triangle defined by the vertex and its
current neighbours.
}
\examples{
set.seed(1)
N <- 10
x <- runif(N)
y <- runif(N)
vis_indices(x, y, 4)
}
