% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvalingam-c.R
\name{vis_simplify}
\alias{vis_simplify}
\title{Line simplification using Visvalingam's algorithm}
\usage{
vis_simplify(x, y, n)
}
\arguments{
\item{x, y}{points}

\item{n}{number of points to keep}
}
\value{
list with elements \code{x} and \code{y} of the simplified line
}
\description{
Visvalingam's algorithm deletes vertices from a line based up the effecive
area of the triangle defined by the vertex and its current neighbours.
}
\details{
The vertex with the minimal area is removed from the line at each iteration
until the target number of points is reached.

Note that when vertices are removed from the line, the effective areas of the
neighbouring vertices need to be recalculated.
}
\examples{
set.seed(1)
N <- 10
x <- runif(N)
y <- runif(N)
vis_simplify(x, y, 4)
}
