% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvalingam-r.R
\name{vis_simplify_r}
\alias{vis_simplify_r}
\title{Naive R implememtation of Visvalingam's line simplification algorithm}
\usage{
vis_simplify_r(x, y, n)
}
\arguments{
\item{x, y}{points}

\item{n}{number of points to keep}
}
\description{
This R code is going to remain unoptimised as a readable demonstration of
the core algorithm.
}
\examples{
set.seed(1)
N <- 10
x <- runif(N)
y <- runif(N)
vis_simplify_r(x, y, 4)
}
